adding earlyState and allowing the beneficiary to withdraaw during that state
we add the early state to the campaign
allow the beneficiary to withdraw the money during this state and
take in to account the money that it had been withdrawn with some state variables
also tso, its posibl to contribute during that state

all thest were passed

+++ contracts/CrowdfyFabric.sol
@@ -59,9 +59,9 @@ contract CrowdfyFabric is CrowdfyFabricI{
     ///@notice deploy a new instance of the campaign
     function createCampaign(
         string calldata _campaignName, 
-        uint _fundingGoal, 
-        uint _deadline, 
-        uint _fundingCap, 
+        uint256 _fundingGoal, 
+        uint256 _deadline, 
+        uint256 _fundingCap, 
         address _beneficiaryAddress
     ) external override returns(uint256) {
